3: ColorImage (Original)
This script will display & manipulate images as per the canvas class in
the comp102x library. To run it, right click on Lab03, select new
Lab03(). On the new instance below, right click on it. On the new menu,
select loadAnImage for task 1, loadTwoImagesOnTheSameCanvas for task 2,
and imageMultiplication for task 3.